requirements.txt
# Core Flask
flask>=3.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
flask-cors>=4.0
flask-limiter>=3.5
gunicorn>=21.2

# Database
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1  # PostgreSQL driver (production)
# SQLite is bundled with Python (dev)

# Configuration
pydantic>=2.5
pydantic-settings>=2.1
python-dotenv>=1.0

# Authentication
python-jose[cryptography]>=3.3  # JWT
passlib[bcrypt]>=1.7  # Password hashing
authlib>=1.3  # OAuth (Google, Kakao)

# Background workers
celery>=5.3
redis>=5.0

# External services
youtube-transcript-api>=0.6.0
requests>=2.31

# Monitoring
sentry-sdk[flask]>=1.40

# Testing
pytest>=8.0
pytest-flask>=1.3
factory-boy>=3.3
httpx>=0.26
